+++ tests/modules/ejector/test_ejector.py
@@ -1,5 +1,5 @@
 from typing import Iterable, cast
-from unittest.mock import Mock, MagicMock
+from unittest.mock import Mock
 
 import pytest
 from web3.exceptions import ContractCustomError
@@ -12,18 +12,10 @@ from src.constants import (
     MAX_SEED_LOOKAHEAD,
     MIN_ACTIVATION_BALANCE,
     MIN_VALIDATOR_WITHDRAWABILITY_DELAY,
-    MAX_WITHDRAWALS_PER_PAYLOAD,
-    ETH1_ADDRESS_WITHDRAWAL_PREFIX,
-    COMPOUNDING_WITHDRAWAL_PREFIX,
 )
 from src.modules.ejector import ejector as ejector_module
 from src.modules.ejector.ejector import (
     Ejector,
-    Withdrawal,
-    _get_sweep_delay_in_epochs_post_pectra,
-    predict_withdrawals_number_in_sweep_cycle,
-    get_pending_partial_withdrawals,
-    get_validators_withdrawals,
 )
 from src.modules.ejector.ejector import logger as ejector_logger
 from src.modules.ejector.types import EjectorProcessingState
@@ -32,11 +24,8 @@ from src.modules.submodules.types import ChainConfig, CurrentFrame
 from src.providers.consensus.types import (
     BeaconStateView,
     Checkpoint,
-    ValidatorState,
-    PendingPartialWithdrawal,
-    Validator,
 )
-from src.types import BlockStamp, Gwei, ReferenceBlockStamp, SlotNumber
+from src.types import BlockStamp, Gwei, ReferenceBlockStamp
 from src.utils import validator_state
 from src.web3py.extensions.contracts import LidoContracts
 from src.web3py.extensions.lido_validators import NodeOperatorId, StakingModuleId
@@ -487,77 +476,6 @@ def test_get_sweep_delay_in_epochs_pre_electra(
         # all 1024 validators
         result = ejector._get_sweep_delay_in_epochs(ReferenceBlockStampFactory.build(slot_number=1))
         assert result == 1, "Unexpected sweep delay in epochs"
-
-
-@pytest.mark.unit
-def test_get_sweep_delay_in_epochs_post_pectra(monkeypatch):
-    # Create mock objects for state and spec
-    state = Mock(spec=BeaconStateView)
-    spec = Mock(spec=ChainConfig)
-    spec.slots_per_epoch = 12
-    predicted_withdrawals = 1000
-    with monkeypatch.context() as m:
-        m.setattr(
-            ejector_module,
-            "predict_withdrawals_number_in_sweep_cycle",
-            Mock(return_value=predicted_withdrawals),
-        )
-        # Calculate delay
-        result = _get_sweep_delay_in_epochs_post_pectra(state, spec)
-
-        # Assert the delay calculation is correct
-        expected_delay = predicted_withdrawals / MAX_WITHDRAWALS_PER_PAYLOAD / int(spec.slots_per_epoch) // 2
-        assert result == expected_delay, f"Expected delay {expected_delay}, got {result}"
-
-
-def create_fake_state(
-    validators: list[Validator], pending_partial_withdrawals: list[PendingPartialWithdrawal]
-) -> BeaconStateView:
-    for i in range(len(validators)):
-        validators[i].index = i
-    beacon_state = MagicMock(spec=BeaconStateView)
-    beacon_state.slot = SlotNumber(1)
-    beacon_state.indexed_validators = validators
-    beacon_state.validators = [v.validator for v in validators]
-    beacon_state.balances = [v.balance for v in validators]
-    beacon_state.pending_partial_withdrawals = pending_partial_withdrawals
-    beacon_state.finalized_checkpoint = Checkpoint(epoch="1", root="0x")
-    return beacon_state
-
-
-@pytest.fixture()
-def fake_beacon_state_view():
-    """Fixture to create a fake BeaconStateView."""
-    validators = [
-        LidoValidatorFactory.build_with_balance(Gwei(1000)),
-        LidoValidatorFactory.build_with_balance(Gwei(MIN_ACTIVATION_BALANCE + 1)),
-        LidoValidatorFactory.build_with_balance(Gwei(MIN_ACTIVATION_BALANCE + 1)),
-    ]
-    pending_partial_withdrawals = [
-        PendingPartialWithdrawal(validator_index="0", amount=500, withdrawable_epoch=1),
-        PendingPartialWithdrawal(validator_index="1", amount=700, withdrawable_epoch=1),
-    ]
-    return create_fake_state(validators, pending_partial_withdrawals)
-
-
-@pytest.mark.unit
-def test_get_pending_partial_withdrawals(fake_beacon_state_view):
-    """Test for the `get_pending_partial_withdrawals` function."""
-    result = get_pending_partial_withdrawals(fake_beacon_state_view)
-    assert len(result) == 1, f"Expected 1 pending partial withdrawals, got {len(result)}"
-
-    assert result[0].validator_index == 1, f"Expected validator_index 1, got {result[0].validator_index}"
-    assert result[0].amount == 1, f"Expected amount 1 for validator 1, got {result[0].amount}"
-
-
-@pytest.mark.unit
-def test_get_validators_withdrawals(fake_beacon_state_view):
-    """Test for the `get_validators_withdrawals` function."""
-    result = get_validators_withdrawals(fake_beacon_state_view, [Withdrawal(validator_index=1, amount=1)])
-    assert len(result) == 1, f"Expected 1 withdrawals, got {len(result)}"
-
-    assert result[0].validator_index == 2, f"Expected validator_index 2, got {result[0].validator_index}"
-    assert result[0].amount == 1, f"Expected amount 1 for validator 2, got {result[0].amount}"
 
 
 @pytest.mark.unit
